requirements.txt
django>=3.2
djangorestframework>=3.12
djangorestframework-simplejwt>=5.0
django-filter>=22.1.3
python-dotenv>=1.0.0
